fix(backend): preserve explicit null error data under PocketBase v0.40 JSON v2
PocketBase v0.40 serializes responses with encoding/json/v2, where
omitempty drops a non-nil *any pointing to a nil interface. Use
omitzero on StructuredError.Data so error envelopes keep the
"data": null field when a handler passes explicit null data.

Also adopts Record.GetInt64 for byte-size record fields and updates
the changelog for the v0.40 update (non-zero CLI exit codes, COOP
security header).

+++ backend/internal/storage/service.go
@@ -164,7 +164,7 @@ func (s *Service) GetMetadata(ctx context.Context, storageID string) (map[string
 		"createdBy":   record.GetString(schema.FieldStorageCreatedBy),
 		"filename":    record.GetString(schema.FieldStorageFilename),
 		"contentType": record.GetString(schema.FieldStorageContentType),
-		"size":        record.GetInt(schema.FieldStorageSize),
+		"size":        record.GetInt64(schema.FieldStorageSize),
 		"sha256":      record.GetString(schema.FieldStorageSha256),
 		"extension":   fileExtension(record.GetString(schema.FieldStorageFilename), record.GetString(schema.FieldStorageContentType)),
 	}
@@ -358,7 +358,7 @@ func (s *Service) Upload(ctx context.Context, token string, body io.Reader, cont
 		return "", &UploadError{Code: ErrorCodeUploadExpired, Message: "upload token expired"}
 	}
 
-	maxSize := int64(tokenRec.GetInt(schema.FieldTokenMaxSize))
+	maxSize := tokenRec.GetInt64(schema.FieldTokenMaxSize)
 	if maxSize <= 0 || maxSize > s.config.MaxFileSize {
 		maxSize = s.config.MaxFileSize
 	}
